Move basic types to their own header file
